// apps/srv/src/modules/dealers/interfaces/csv-dump-entry.interface.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface CsvDumpEntryInterface {
    Retail_Date: string;
    Delaer_HO_code: string;
    Name_of_HeadOffice: string;
    Branch_code: string;
    Branch_Name: string;
    Car_Line_Desc: string;
    Car_Desc: string;
    VIN: string;
    City: string;
    MobiKeyActivateDate: string;
    MobiKeyStatus: string;
    MobiKeyType: string;
}